Separate class variants fns for hot reload
