+++ src/components/ui/skeleton.tsx
@@ -1,7 +1,10 @@
 import { cn } from '@/lib/utils'
 import * as react from 'react'
 
-function Skeleton({ className, ...props }: react.HTMLAttributes<HTMLDivElement>) {
+function Skeleton({
+	className,
+	...props
+}: react.HTMLAttributes<HTMLDivElement>) {
 	return (
 		<div
 			className={cn('animate-pulse rounded-md bg-muted', className)}
