+++ src/routes/_user/profile.index.tsx
@@ -1,4 +1,4 @@
-import { buttonVariants } from '@/components/ui/button'
+import { buttonVariants } from '@/components/ui/button-variants'
 import {
 	Card,
 	CardContent,
