+++ src/routes/learn/add-deck.tsx
@@ -4,7 +4,8 @@ import { z } from 'zod'
 import { zodResolver } from '@hookform/resolvers/zod'
 import { ShowError } from '@/components/errors'
 import { SelectOneLanguage } from '@/components/select-one-language'
-import { Button, buttonVariants } from '@/components/ui/button'
+import { Button } from '@/components/ui/button'
+import { buttonVariants } from '@/components/ui/button-variants'
 import { useNewDeckMutation } from '@/lib/mutate-deck'
 import { useProfile } from '@/lib/use-profile'
 import { NavbarData } from '@/types/main'
